requirements-dev.txt
WebTest==3.0.0
django-webtest==1.9.8
mkdocs==1.2.3
Babel==2.9.1
pytest-django==4.5.2
pytest-cov==3.0.0
